requirements.txt
gunicorn==20.0.4
flask-socketio==4.3.1
absl-py==0.9.0
aniso8601==8.0.0
astor==0.8.1
astroid==2.3.3
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
gast==0.3.3
#google-auth==1.11.0
# google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8

google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0

grpcio==1.26.0
h5py==2.10.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.0
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.0
Pillow==7.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.4.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.1
eventlet==0.29.1
gevent~=20.9.0